% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montserrat.R
\docType{data}
\name{font_pub}
\alias{font_pub}
\alias{font_mon_bold}
\alias{font_mon_light}
\alias{font_mon_thin}
\title{Montserrat font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_pub

font_mon_bold

font_mon_light

font_mon_thin
}
\description{
\code{font_pub} == "\verb{Public Sans}"

\code{font_mon_bold} == "\verb{Montserrat Bold}"

\code{font_mon_light} == "\verb{Montserrat Light}"

\code{font_mon_thin} == "\verb{Montserrat Thin}"
}
\note{
\code{font_mon_bold} (a.k.a. "\verb{Montserrat Bold}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
