% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_projections}
\alias{population_projections}
\title{`population_projections` is a `tibble` that contains data from the Texas
Demographic Center, run by the State's Official Demographer,  Dr. Lloyd Potter.}
\format{
A data frame:
}
\usage{
population_projections
}
\description{
`population_projections` is a `tibble` that contains data from the Texas
Demographic Center, run by the State's Official Demographer,  Dr. Lloyd Potter.
}
\keyword{datasets}
