% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_hc.R
\name{tx2036_hc_light}
\alias{tx2036_hc_light}
\title{Texas 2036 highcharter Themes}
\usage{
tx2036_hc_light(
  base_size = "12px",
  base_family = "Montserrat",
  title_size = "28px",
  title_weight = "800",
  subtitle_size = "14px",
  subtitle_weight = "400",
  credit_size = "10px",
  header_align = "left",
  legend_align = "center",
  credits_align = "left",
  ...
)
}
\arguments{
\item{base_size}{the base font size of text}

\item{base_family}{the font family}

\item{title_size}{the size of your title font}

\item{title_weight}{the weight your tiitle font}

\item{subtitle_size}{the size your subtitle font}

\item{subtitle_weight}{the size your subtitle font}

\item{credit_size}{the size your credits font}

\item{header_align}{where to align the title and subtitle}

\item{legend_align}{where to align the legend}

\item{credits_align}{where to align the credits}

\item{...}{Additional arguments passed to [highcharter::hc_theme()]}
}
\value{
a highcharter theme
}
\description{
Minimalistic highcharter themes for use on Texas 2036 sites.
}
\examples{

data(economics_long, package = "ggplot2")

economics_long2 <- dplyr::filter(economics_long, variable \%in\% c("pop", "uempmed", "unemploy"))

head(economics_long2)

hchart(economics_long2, "line", hcaes(x = date, y = value01, group = variable)) \%>\%
tx2036_hc_light()
}
