% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_projections}
\alias{water_projections}
\title{`water_projections` is a `tibble` that contains data from the Texas
Water Development Board.}
\format{
A data frame:
}
\usage{
water_projections
}
\description{
`water_projections` is a `tibble` that contains data from the Texas
Water Development Board.
}
\keyword{datasets}
